[REF/IMP] global: improves project structure and encapsulates parser logic
