+++ api/.gitignore
@@ -2,3 +2,6 @@
 *.log
 .venv/
 notebooks/
+archive/
+app.log
+__pycache__/
