+++ api/configs.py
@@ -1,0 +1,14 @@
+import sys
+from pathlib import Path
+
+BASE_DIR = Path(__file__).resolve()
+sys.path.append(str(BASE_DIR / "src"))
+
+FRONTEND_ORIGIN = "http://localhost:8080"
+
+class config:
+    origins = [
+        FRONTEND_ORIGIN,
+    ]
+    app_name = "reader"
+
