--- api/requirements/all.txt
@@ -1,145 +1,0 @@
-aiohappyeyeballs==2.4.6
-aiohttp==3.11.12
-aiosignal==1.3.2
-annotated-types==0.7.0
-anyio==4.8.0
-appnope==0.1.4
-asttokens==3.0.0
-attrs==25.1.0
-beautifulsoup4==4.13.3
-cachetools==5.5.1
-certifi==2025.1.31
-cffi==1.17.1
-charset-normalizer==3.4.1
-click==8.1.8
-comm==0.2.2
-cryptography==44.0.0
-debugpy==1.8.12
-decorator==5.1.1
-deprecation==2.1.0
-distro==1.9.0
-dnspython==2.7.0
-email_validator==2.2.0
-executing==2.2.0
-fastapi==0.115.8
-fastapi-cli==0.0.7
-filelock==3.17.0
-filetype==1.2.0
-frozenlist==1.5.0
-fsspec==2025.2.0
-ftfy==6.3.1
-google-ai-generativelanguage==0.6.15
-google-api-core==2.24.1
-google-api-python-client==2.160.0
-google-auth==2.38.0
-google-auth-httplib2==0.2.0
-google-generativeai==0.8.4
-googleapis-common-protos==1.67.0rc1
-gotrue==2.11.3
-grpcio==1.70.0
-grpcio-status==1.70.0
-h11==0.14.0
-h2==4.2.0
-hpack==4.1.0
-httpcore==1.0.7
-httplib2==0.22.0
-httptools==0.6.4
-httpx==0.28.1
-huggingface-hub==0.28.1
-hyperframe==6.1.0
-idna==3.10
-ipykernel==6.29.5
-ipython==8.32.0
-jedi==0.19.2
-Jinja2==3.1.5
-jiter==0.8.2
-joblib==1.4.2
-jsons==1.6.3
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-markdown-it-py==3.0.0
-markdown2==2.5.3
-markdownify==0.13.1
-marker-pdf==1.3.5
-MarkupSafe==3.0.2
-matplotlib-inline==0.1.7
-mdurl==0.1.2
-mpmath==1.3.0
-multidict==6.1.0
-nest-asyncio==1.6.0
-networkx==3.4.2
-numpy==2.2.2
-openai==1.61.1
-opencv-python==4.11.0.86
-packaging==24.2
-parso==0.8.4
-pdfminer.six==20240706
-pdftext==0.5.1
-pexpect==4.9.0
-pillow==10.4.0
-platformdirs==4.3.6
-postgrest==0.19.3
-prompt_toolkit==3.0.50
-propcache==0.2.1
-proto-plus==1.26.0
-protobuf==5.29.3
-psutil==6.1.1
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pyasn1==0.6.1
-pyasn1_modules==0.4.1
-pycparser==2.22
-pydantic==2.10.6
-pydantic-settings==2.7.1
-pydantic_core==2.27.2
-Pygments==2.19.1
-pyparsing==3.2.1
-pypdfium2==4.30.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-python-multipart==0.0.20
-PyYAML==6.0.2
-pyzmq==26.2.1
-RapidFuzz==3.12.1
-realtime==2.3.0
-regex==2024.11.6
-requests==2.32.3
-rich==13.9.4
-rich-toolkit==0.13.2
-rsa==4.9
-safetensors==0.5.2
-scikit-learn==1.6.1
-scipy==1.15.1
-shellingham==1.5.4
-six==1.17.0
-sniffio==1.3.1
-soupsieve==2.6
-stack-data==0.6.3
-starlette==0.45.3
-storage3==0.11.3
-StrEnum==0.4.15
-supabase==2.13.0
-supafunc==0.9.3
-surya-ocr==0.10.3
-sympy==1.13.1
-termcolor==2.5.0
-texify==0.2.1
-threadpoolctl==3.5.0
-tokenizers==0.21.0
-torch==2.6.0
-torchvision==0.21.0
-tornado==6.4.2
-tqdm==4.67.1
-traitlets==5.14.3
-transformers==4.48.3
-typer==0.15.1
-typing_extensions==4.12.2
-typish==1.9.3
-uritemplate==4.1.1
-urllib3==2.3.0
-uvicorn==0.34.0
-uvloop==0.21.0
-watchfiles==1.0.4
-wcwidth==0.2.13
-websockets==14.2
-yarl==1.18.3
