+++ api/src/books/constants.py
@@ -1,0 +1,1 @@
+MAX_FILE_SIZE = 10 * 1024 * 1024
