+++ api/src/books/schemas.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel, Field
+
+class Response(BaseModel):
+    message: str = Field(description="Response message")
